added modal for new task
take in user input for adding new tasks to do. just shell
